kernel: move main
main was in lib.rs, which was a little tricky to find. Also do_process
in sched.rs had to be public, but is only called in one place. This
merges the two, so all scheduling and main loop code is in the same
place, and renames the file to main.rs so it is obvious where everything
starts.